restructured the backend files to use one folder
